♻️ refactor(widgets): split sections into presentational widgets
