+++ src/features/theme/model/theme-atoms.ts
@@ -1,7 +1,7 @@
 import { atom } from "jotai";
 import { atomWithStorage, createJSONStorage } from "jotai/utils";
 
-import { storageKey } from "../../../lib/constants";
+import { storageKey } from "../../../shared/lib";
 import { getSystemTheme, resolveTheme, type ResolvedTheme, type ThemeMode } from "../lib/theme";
 
 const themeStorage = createJSONStorage<ThemeMode>(() => localStorage);
