--- src/sections/about/about-section.stories.tsx
@@ -1,17 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react-vite";
-
-import { AboutSection } from "./about-section";
-
-const meta = {
-  component: AboutSection,
-  parameters: {
-    layout: "fullscreen",
-  },
-  title: "Sections/AboutSection",
-} satisfies Meta<typeof AboutSection>;
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {};
