--- src/sections/contact/contact-section.stories.tsx
@@ -1,17 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react-vite";
-
-import { ContactSection } from "./contact-section";
-
-const meta = {
-  component: ContactSection,
-  parameters: {
-    layout: "fullscreen",
-  },
-  title: "Sections/ContactSection",
-} satisfies Meta<typeof ContactSection>;
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {};
