+++ src/widgets/about/index.ts
@@ -1,0 +1,1 @@
+export { AboutSection } from "./ui/about-section";
