+++ src/widgets/about/ui/about-section.stories.tsx
@@ -1,0 +1,22 @@
+import type { Meta, StoryObj } from "@storybook/react-vite";
+
+import { techStacks } from "../../../entities/profile";
+import { AboutSectionUI } from "./about-section.ui";
+
+const meta = {
+  component: AboutSectionUI,
+  parameters: {
+    layout: "fullscreen",
+  },
+  title: "Widgets/AboutSection",
+} satisfies Meta<typeof AboutSectionUI>;
+
+export default meta;
+
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  args: {
+    techStacks,
+  },
+};
