+++ src/widgets/about/ui/about-section.tsx
@@ -1,0 +1,6 @@
+import { techStacks } from "../../../entities/profile";
+import { AboutSectionUI } from "./about-section.ui";
+
+export function AboutSection() {
+  return <AboutSectionUI techStacks={techStacks} />;
+}
