+++ src/widgets/contact/index.ts
@@ -1,0 +1,1 @@
+export { ContactSection } from "./ui/contact-section";
