+++ src/widgets/contact/ui/contact-section.tsx
@@ -1,0 +1,6 @@
+import { SocialLinks } from "../../../entities/profile";
+import { ContactSectionUI } from "./contact-section.ui";
+
+export function ContactSection() {
+  return <ContactSectionUI slots={{ SocialLinks: <SocialLinks /> }} />;
+}
