+++ src/widgets/floating-navigation/index.ts
