+++ src/widgets/floating-navigation/model/panel-state.ts
