+++ src/widgets/floating-navigation/model/use-floating-navigation-panel.ts
