+++ src/widgets/floating-navigation/ui/floating-navigation.stories.tsx
