+++ src/widgets/floating-navigation/ui/panels/parts.test.tsx
