+++ src/widgets/footer/ui/footer-section.tsx
@@ -1,0 +1,5 @@
+import { FooterSectionUI } from "./footer-section.ui";
+
+export function FooterSection() {
+  return <FooterSectionUI />;
+}
