+++ src/widgets/footer/ui/footer-section.ui.tsx
@@ -1,0 +1,10 @@
+export function FooterSectionUI() {
+  return (
+    <footer className="h-25 bg-slate-1 px-6">
+      <div className="mx-auto flex h-full w-full max-w-160 items-center justify-between gap-4 text-center text-xs text-slate-11">
+        <p className="font-medium">&copy; 2026 koutyuke</p>
+        <p className="font-handwritten">Thanks for visiting!</p>
+      </div>
+    </footer>
+  );
+}
