+++ src/widgets/footprints/index.ts
@@ -1,0 +1,1 @@
+export { FootprintsSection } from "./ui/footprints-section";
