+++ src/widgets/footprints/ui/footprints-section.tsx
@@ -1,0 +1,6 @@
+import { footprints } from "../../../entities/footprint";
+import { FootprintsSectionUI } from "./footprints-section.ui";
+
+export function FootprintsSection() {
+  return <FootprintsSectionUI footprints={footprints} />;
+}
