+++ src/widgets/hero/index.ts
@@ -1,0 +1,1 @@
+export { HeroSection } from "./ui/hero-section";
